# Store bounding box sizes in half precision