I finally found a use for half precision, though it's just for storage, not computation.  This reduces memory bandwidth when building the neighbor list.  For very large systems it can produce a bit of a speedup.  For smaller systems the effect is negligible.  See https://github.com/openmm/openmm/issues/4063#issuecomment-1537514168.

cc @philipturner