@philipturner 2023-05-08
The Apple GPU excels in a lot of the control flow operations present in the $O(n^2)$ part. That means removing a memory bandwidth bottleneck should cause a greater % improvement than with NV.

With GPT-4, it might be possible for me to convert the OpenMM Metal plugin into actual Metal shaders. Can we isolate the GPU shader for `findBlocksWithInteractions` so I can run it through Metal in a Swift script? I need all but the absolutely necessary shader arguments to be elided.