@peastman 2023-05-08
> Can we isolate the GPU shader for findBlocksWithInteractions so I can run it through Metal in a Swift script?

I'm not sure what you're asking.  Can you clarify?

The changes here are mostly trivial, just converting floats to half before saving them and converting back when loading.  The extra code at the top is just to work around some CUDA annoyances.