@philipturner 2023-05-09
I think I found a workaround that works for OpenCL! This compiled in the Apple OpenCL driver. I tried it after experiencing the following error:

```
(base) philipturner@M1-Max-MacBook-Pro CompileCL % xcrun metal -c File.cl --driver-mode=openclc -o File.air
File.cl:7:17: error: loading directly from pointer to type '__global half' requires
      cl_khr_fp16. Use vector data load builtin functions instead
  if (*input1 > *input2) {
                ^
1 error generated.
(base) philipturner@M1-Max-MacBook-Pro CompileCL % vim File.cl 
(base) philipturner@M1-Max-MacBook-Pro CompileCL % xcrun metal -c File.cl --driver-mode=openclc -o File.air
```

float vload_half(size_t offset, const __global half *p) float vload_half(size_t offset, const __local half *p) float vload_half(size_t offset, const __constant half *p) float vload_half(size_t offset, const __private half *p)  								For OpenCL C 2.0, or OpenCL C 3.0 or newer with the __opencl_c_ generic_address_space feature:  								float vload_half(size_t offset, const half *p) | Read sizeof(half) bytes of data from the address computed as (p + offset). The data read is interpreted as a half value. The half value is converted to a float value and the float value is returned. The computed read address must be 16-bit aligned.
-- | --
floatn vload_halfn(size_t offset, const __global half *p) floatn vload_halfn(size_t offset, const __local half *p) floatn vload_halfn(size_t offset, const __constant half *p) floatn vload_halfn(size_t offset, const __private half *p)  								For OpenCL C 2.0, or OpenCL C 3.0 or newer with the __opencl_c_ generic_address_space feature:  								floatn vload_halfn(size_t offset, const half *p) | Read (n * sizeof(half)) bytes of data from the address computed as (p + (offset * n)). The data read is interpreted as a halfn value. The halfn value read is converted to a floatn value and the floatn value is returned. The computed read address must be 16-bit aligned.

OpenCL code:

```opencl
                                                                                 
__kernel void testKernel(__global float* input1,                                 
                         __global half* input2,                                  
                         __global int* output)                                   
{                                                                                
  if (*input1 > vload_half(0, input2)) {                                         
    output[0] = 1;                                                               
  }                                                                              
}  
```