@philipturner 2023-05-09
Optimization 1: `real4` -> `real3` for positions and box sizes

Optimization 2: `real3` -> `half` for box sizes

Width (nm) | Atoms | Steps/4 | Original | Opt. 1 | Opt. 2 | Opt. 2 Time / Before
-- | -- | -- | -- | -- | -- | --
2 | 774 | 1000 | 889 | 893 | 889 | 1.004
3 | 2661 | 1000 | 668 | 668 | 669 | 0.999
4 | 6282 | 1000 | 506 | 505 | 505 | 1.000
5 | 12255 | 1000 | 328 | 328 | 327 | 1.003
7 | 33840 | 1000 | 220 | 220 | 220 | 1.000
10 | 98880 | 1000 | 88.9 | 88.9 | 88.8 | 1.001
15 | 335025 | 100 | 26.4 | 26.7 | 26.7 | 1.000
20 | 792450 | 50 | 11.1 | 11.0 | 11.3 | 0.982
25 | 1550160 | 25 | 5.01 | 5.03 | 5.18 | 0.971
30 | 2682600 | 15 |   | 2.38 | 2.51 | 0.948
35 | 4254045 | 10 | 1.20 | 1.14 | 1.30 | 0.923
40 | 6354990 | 10 | 0.692 |   | 0.763 | 0.907
45 | 9056475 | 5 | 0.369 |   | 0.443 | 0.833
apoa1rf | 92224 | 6800 | 151.7 | 151.6 | 151.7 | 1.000

https://github.com/philipturner/openmm-metal/commit/0b487005873c3bbb1c40520fe6cc4fd1b198f0b4