@peastman 2023-05-09
Does that mean `vload_half()` and `vstore_half()` are available even without the extension?  I'm looking at the OpenCL 1.2 spec, and it's a little unclear.  There's a footnote in section 6.12.7 that says, "The half*n* type is only defined by the **cl_khr_fp16** extension described in section 9.5 of the OpenCL 1.2 Extension Specification."  Does that include the half type itself, or do only vectors of halfs require the extension?

Anyway, if it works on all platforms, great!