@peastman 2023-05-09
I was able to implement half precision support in OpenCL.  It works, but it has no measurable effect on speed on either NVIDIA or Apple.

|Width|Atoms|Old|New|Ratio|
|---|---|---|---|---|
15|335025|10.03|9.97|0.99
20|792450|29.67|29.79|1.00
25|1550160|78.21|78.23|1.00

Here's my implementation.

```c
real3 loadHalf3(const __global half* h, size_t i) {
    return (real3) (vload_half(3*i, h), vload_half(3*i+1, h), vload_half(3*i+2, h));
}

void storeHalf3(real3 f, __global half* h, size_t i) {
    // Round up so we'll err on the side of making the box a little too large.
    // This ensures interactions will never be missed.
    vstore_half_rtp((float) f.x, 3*i, h);
    vstore_half_rtp((float) f.y, 3*i+1, h);
    vstore_half_rtp((float) f.z, 3*i+2, h);
}
```

Possibly the problem is that it has to load the three components as three separate 16 bit memory transactions, rather than a single 48 bit one.  Loading all three at once requires the fp16 extension.