@philipturner 2023-05-09
> Loading all three at once requires the fp16 extension.

The OpenCL C specification does not state any such requirement for this function. The function does not use the `half3` type.

| Function | Description |
| ---- | ---- |
| floatn vloada_halfn(size_t offset, const __global half *p) floatn vloada_halfn(size_t offset, const __local half *p) floatn vloada_halfn(size_t offset, const __constant half *p) floatn vloada_halfn(size_t offset, const __private half *p) | For n = 2, 4, 8 and 16, read sizeof(halfn) bytes of data from the address computed as (p + (offset * n)). The data read is interpreted as a halfn value. The halfn value read is converted to a floatn value and the floatn value is returned. The computed address must be aligned to sizeof(halfn) bytes. For n = 3, vloada_half3 reads a half3 from the address computed as (p + (offset * 4)) and returns a float3. The computed address must be aligned to sizeof(half) * 4 bytes.

```opencl
__global half* restrict sortedBlockBoundingBox,
vstorea_half3_rtp(float3(blockBoundingBox[index]), i, sortedBlockBoundingBox);

__global const half* restrict sortedBlockBoundingBox,
real3 blockSizeY = real3(vloada_half3(block2, sortedBlockBoundingBox));
```

Works on Apple, which does not have `cl_khr_fp16`:

```
  printf() buffer size                            1048576 (1024KiB)
  Built-in kernels                                (n/a)
  Device Extensions                               cl_APPLE_SetMemObjectDestructor cl_APPLE_ContextLoggingFunctions cl_APPLE_clut cl_APPLE_query_kernel_names cl_APPLE_gl_sharing cl_khr_gl_event cl_khr_byte_addressable_store cl_khr_global_int32_base_atomics cl_khr_global_int32_extended_atomics cl_khr_local_int32_base_atomics cl_khr_local_int32_extended_atomics cl_khr_3d_image_writes cl_khr_image2d_from_buffer cl_khr_depth_images 
```